Refactor: 모바일 메인페이지의 헤더(로그인, 쿠폰함이동)를 메인 컴포넌트에서 헤더컴포넌트로 이동
