+++ src/pages/Home/Main/Top.js
@@ -5,98 +5,6 @@ import { LoginContext } from "../../../contexts/LoginContextProvider";
 
 /** Top 15 또또가 */
 const Top = ({ top15 }) => {
-  // const top15 = [
-  //   {
-  //     id: 1,
-  //     rank: 1,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 2,
-  //     rank: 2,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 3,
-  //     rank: 3,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 4,
-  //     rank: 4,
-  //     text: "[유-엠씨] 글씨가",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 5,
-  //     rank: 5,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 6,
-  //     rank: 6,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 7,
-  //     rank: 7,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 8,
-  //     rank: 8,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 9,
-  //     rank: 9,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 10,
-  //     rank: 10,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 11,
-  //     rank: 11,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 12,
-  //     rank: 12,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 13,
-  //     rank: 13,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 14,
-  //     rank: 14,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  //   {
-  //     id: 15,
-  //     rank: 15,
-  //     text: "[유-엠씨] 글씨가 매우매우 매우매우 매우매우매우매우매우 길어진다면 이렇게 말줄임표를 써서 보여주세요",
-  //     reviewCount: 65,
-  //   },
-  // ];
   return (
     <div className="my-16">
       {/* 구분선 */}
